Add test for unit structs, and fix existing test.
Serializing unit structs as arrays is fundamentally incompatible with
how we currently deserialize them. We need to either A) change the
deserializer to care about whether deserialize_seq or deserialize_unit
was used to change our interpretation of an empty sequence, or B) never
serialize unit structs as empty sequences. B seems easier?