Serialize seqs and maps of unknown length (#196)
This fixes the serialization of structs that contain fields tagged with #[serde(flatten)].
The basic idea is to write elements to another Serializer, count the elements that are
serialized, and in the end, write the counted length and the serialzed elements to the
main Serializer.

A drawback might be that such a struct cannot be written in compact representation as
serde wants to serialize it as a map.