Experimental: Support no_std encoding for ByteBuf
Because std::io::Write is missing from the `core` crate,
this introduces our own `RmpWrite` trait (and a ByteBuf type).

For no_std targets, the only implementation of RmpWrite is for our
own ByteBuf type (inspired by serde_bytes).

When #[cfg(feature="std")] is enabled (the default), then
RmpWrite has a blanket implementation for std::io::Write.
This makes it is 100% backwards compatible.

Because of this, introducing this trait required 0 changes to the tests or the dependent crates.

Because of this careful backwards compatibility, these changes should only
require a patch version bump for the `rmp` crate (we are major version 0).

This also moves several (internal) utility methods to RmpWrite.
Further refactoring should be possible in the future.

TODO: Implement no_std decoding (that shouldn't even require the alloc crate)