Cleanup: Update internal APIs in rmpv/rmp-serde
The internal read_data_* functions moved from the top level to
the RmpRead trait as part of supporting no_std.

Because of compatibility shims, using the old functions still works.
That means this commit is *not* required to get anything to compile.

Tests for rmpv & rmp-serde compile
and run both before and after this commit.

However, using the old code (with the compatibility shims) now comes with deprecation warnings.

This commit fixes those deprecation warnings, by using the new functions
located in the RmpRead trait.

However, this means the new (cleaned up) code does not work without the RmpRead trait.
As such, we have to bump the version of the internal 'rmpv' dependency in both these crates.

Once again, this commit isn't required to get any code to compile/work.
All tests compile and run both before and afterwards.
It is just needed to fix some deprecation warnings.